refactor: 클릭 이동 추가 및 import문 변경
