+++ src/components/myfamily/TravelRecord.tsx
@@ -2,7 +2,6 @@
 
 import React from 'react';
 import { useRouter } from 'next/navigation';
-
 import Image from 'next/image';
 import classNames from 'classnames/bind';
 import styles from './TravelRecord.module.scss';
